RAG-Based-Multi-Source-Chatbot-Using-LLM-main/requirements.txt
langchain_community
langchain
faiss-cpu
sentence-transformers==2.4.0
InstructorEmbedding
pypdf
streamlit
pynvml
torch
python-dotenv
